Rename sort by key for person search
