+++ src/pages/search/project_search/ProjectSearch.tsx
@@ -2,6 +2,7 @@ import { Box, List, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { useHistory } from 'react-router-dom';
 import { ListSkeleton } from '../../../components/ListSkeleton';
+import { SortSelector } from '../../../components/SortSelector';
 import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
 import { SearchParam } from '../../../utils/searchHelpers';
 import { CristinSearchPagination } from '../CristinSearchPagination';
@@ -24,13 +25,39 @@ export const ProjectSearch = ({ projectQuery }: ProjectSearchProps) => {
 
   const rowsPerPage = resultsParam ? +resultsParam : ROWS_PER_PAGE_OPTIONS[0];
 
+  const sortingComponent = (
+    <SortSelector
+      orderKey="orderBy"
+      sortKey="sort"
+      aria-label={t('search.sort_by')}
+      size="small"
+      variant="standard"
+      options={[
+        {
+          orderBy: 'name',
+          sortOrder: 'asc',
+          label: t('search.sort_by_name_asc'),
+        },
+        {
+          orderBy: 'name',
+          sortOrder: 'desc',
+          label: t('search.sort_by_name_desc'),
+        },
+      ]}
+    />
+  );
+
   return (
     <Box sx={{ display: 'flex', flexDirection: 'column', gap: '1rem' }}>
       {projectQuery.isLoading ? (
         <ListSkeleton arrayLength={3} minWidth={40} height={100} />
       ) : projectsSearchResults && projectsSearchResults.length > 0 ? (
         <div>
-          <CristinSearchPagination totalCount={totalHits} page={page} rowsPerPage={rowsPerPage}>
+          <CristinSearchPagination
+            totalCount={totalHits}
+            page={page}
+            rowsPerPage={rowsPerPage}
+            sortingComponent={sortingComponent}>
             <List>
               {projectsSearchResults.map((project) => (
                 <ProjectListItem key={project.id} project={project} />
