+++ src/translations/nbTranslations.json
@@ -1764,6 +1764,8 @@
     "select_one_or_more_categories": "Velg en eller flere kategorier",
     "sort_by": "Sorter etter",
     "sort_by_modified_date": "Sist endret",
+    "sort_by_last_name_asc": "Etternavn A-Å",
+    "sort_by_last_name_desc": "Etternavn Å-A",
     "sort_by_name_asc": "Navn A-Å",
     "sort_by_name_desc": "Navn Å-A",
     "sort_by_published_date_asc": "Publiseringsdato (eldste først)",
